refactor components for better separation of components
