+++ src/components/Panels/CoachingStatusPanel/CoachingStatusPanel.Style.js
@@ -1,0 +1,13 @@
+import styled from '@emotion/styled';
+
+export const Status = styled('div')`
+  font-size: 14px;
+  font-weight: bold;
+  margin-top: 10px;
+  margin-bottom: 4px;
+  white-space: nowrap;
+  overflow: hidden;
+  text-overflow: ellipsis;
+  align-items: center;
+  text-align: center;
+`;
