+++ src/components/Panels/SupervisorMonitorPanel/SupervisorMonitorPanel.Style.js
@@ -1,0 +1,12 @@
+import styled from 'react-emotion';
+
+export const Status = styled('div')`
+  font-size: 14px;
+  margin-top: 10px;
+  margin-bottom: 4px;
+  white-space: nowrap;
+  overflow: hidden;
+  text-overflow: ellipsis;
+  align-items: left;
+  text-align: left;
+`;
