+++ src/components/Panels/index.js
@@ -1,0 +1,2 @@
+export { default as CoachingStatusPanel } from './CoachingStatusPanel';
+export { default as SupervisorMonitorPanel } from './SupervisorMonitorPanel';
