--- src/components/SupervisorMonitorPanel.js
@@ -1,89 +1,0 @@
-import * as React from 'react';
-import { withTheme } from '@twilio/flex-ui';
-import styled from 'react-emotion';
-import { connect } from 'react-redux';
-import { bindActionCreators } from 'redux';
-
-import { Actions as BargeCoachStatusAction } from '../states/BargeCoachState';
-import AbstractPanel from './AbstractPanel';
-
-const Status = styled('div')`
-  font-size: 14px;
-  margin-top: 10px;
-  margin-bottom: 4px;
-  white-space: nowrap;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  align-items: left;
-  text-align: left;
-`;
-
-class SupervisorMonitorPanel extends AbstractPanel {
-  componentDidUpdate = async () => {
-    // Setup the listener if it hasn't already and we have an agentSid
-    if (!this.doc && this.props.agentWorkerSID) {
-      await this.setupListener(`syncDoc.${this.props.agentWorkerSID}`);
-    }
-  };
-
-  render() {
-    return (
-      <Status>
-        <div>
-          <h1 id="title" fontWeight="bold">
-            Active Supervisors:
-          </h1>
-          {this.renderSupervisors()}
-        </div>
-      </Status>
-    );
-  }
-
-  renderSupervisors() {
-    if (this.props.supervisorArray.length === 0) {
-      return 'None';
-    }
-
-    return (
-      <table id="supervisors">
-        <tbody>
-          {this.props.supervisorArray.map((supervisorArray) => (
-            <tr key={supervisorArray.supervisor}>
-              <td>{supervisorArray.supervisor}</td>
-              <td style={{ color: 'green' }}>&nbsp;{supervisorArray.status}</td>
-            </tr>
-          ))}
-        </tbody>
-      </table>
-    );
-  }
-}
-
-// Mapping the agent's sid and supervisor full name
-const mapStateToProps = (state) => {
-  const agentWorkerSID = state?.flex?.supervisor?.stickyWorker?.worker?.sid;
-  const supervisorFN = state?.flex?.worker?.attributes?.full_name;
-
-  /*
-   * Also pulling back the states from the redux store as we will use those later
-   * to manipulate the buttons
-   */
-  const customReduxStore = state?.['barge-coach'].bargecoach;
-  const { supervisorArray } = customReduxStore;
-
-  return {
-    agentWorkerSID,
-    supervisorFN,
-    supervisorArray,
-  };
-};
-
-/*
- * Mapping dispatch to props as I will leverage the setBargeCoachStatus
- * to change the properties on the redux store, referenced above with this.props.setBargeCoachStatus
- */
-const mapDispatchToProps = (dispatch) => ({
-  setBargeCoachStatus: bindActionCreators(BargeCoachStatusAction.setBargeCoachStatus, dispatch),
-});
-
-export default connect(mapStateToProps, mapDispatchToProps)(withTheme(SupervisorMonitorPanel));
