+++ src/components/index.js
@@ -1,2 +1,3 @@
 export { default as SupervisorPrivateModeButton } from './SupervisorPrivateModeButton';
 export { default as SupervisorBargeCoachButton } from './SupervisorBargeCoachButton';
+export { CoachingStatusPanel, SupervisorMonitorPanel } from './Panels';
